Extract stuff and move it around
