+++ app/models/parsers/decent_tcl.rb
@@ -2,6 +2,8 @@
 
 module Parsers
   class DecentTcl < Base
+    DATA_LABELS = %w[espresso_pressure espresso_weight espresso_flow espresso_flow_weight espresso_temperature_basket espresso_temperature_mix espresso_water_dispensed espresso_temperature_goal espresso_flow_weight_raw espresso_pressure_goal espresso_flow_goal espresso_state_change].freeze
+
     def initialize(file)
       super
       @start_chars_to_ignore = %i[c b]
@@ -30,7 +32,7 @@ module Parsers
 
     private
 
-    Shot::DATA_LABELS.each do |name|
+    DATA_LABELS.each do |name|
       define_method("extract_#{name}") do |data|
         @data[name] = data
       end
